refactor: session 패키지로 묶음
