+++ tomcat/src/main/java/nextstep/jwp/presentation/RegisterController.java
@@ -1,11 +1,11 @@
 package nextstep.jwp.presentation;
 
-import static org.apache.catalina.Session.JSESSIONID;
+import static org.apache.catalina.session.Session.JSESSIONID;
 import static org.apache.coyote.http11.util.StaticResource.INDEX_PATH;
 
 import nextstep.jwp.model.User;
 import nextstep.jwp.service.UserService;
-import org.apache.catalina.Session;
+import org.apache.catalina.session.Session;
 import org.apache.coyote.http11.common.HttpRequest;
 import org.apache.coyote.http11.common.HttpResponse;
 import org.apache.coyote.http11.constant.HttpStatus;
