+++ tomcat/src/main/java/org/apache/catalina/session/SessionManager.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package org.apache.catalina.session;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
